Proyecto Ituran/Proyecto Ituran/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
et_xmlfile==2.0.0
fastapi==0.121.2
google-api-core==2.28.1
google-auth==2.42.1
google-cloud-bigquery==3.38.0
google-cloud-core==2.5.0
google-cloud-storage==3.5.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.71.0
grpcio==1.76.0
grpcio-status==1.76.0
idna==3.11
numpy==2.3.3
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
proto-plus==1.26.1
protobuf==6.33.0
pyarrow==23.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.4
pydantic_core==2.41.5
pyodbc==5.3.0
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.49.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
